styles conclusions

+++ src/controllers/videoController.js
@@ -1,12 +1,13 @@
-import Video from "../models/Video"
+import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
-    const videos = await Video.find({}).sort({ createdAt: "desc" });
+    const videos = await Video.find({}).sort({ createdAt: "desc" }).populate("owner");
     return res.render("home", { pageTitle: "Home", videos });
 };
 export const watch = async (req, res) =>{
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video) {
         return res.render("404", { pageTitle: "Video not found." });
     }
@@ -14,18 +15,30 @@ export const watch = async (req, res) =>{
 };
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: {_id}
+    } = req.session;
     const video = await Video.findById(id);
     if(!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: {_id}
+    } = req.session;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title, description, hashtags: Video.formatHashtags(hashtags),
@@ -38,23 +51,44 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const {
+        user: { _id },
+    } = req.session;
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
-    const video = new Video ({
-        title: title,
-        description: description,
-        createdAt: Date.now(),
-        hashtags: Video.formatHashtags(hashtags),
-        meta: {
-            views: 0,
-            rating: 0,
-        },
-    });
-    await video.save();
-    return res.redirect("/");
+    try {
+        const newVideo = await Video.create({
+            title,
+            description,
+            fileUrl,
+            owner: _id,
+            hashtags: Video.formatHashtags(hashtags),
+        });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
+        return res.redirect("/");
+    } catch (error) {
+        console.log(error);
+        return res.status(400).render("upload", {
+            pageTitle: "Upload Video",
+            errorMessage: error._message,
+        });
+    }
 };
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: {_id}
+    } = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
@@ -65,9 +99,9 @@ export const search = async (req, res) => {
     if (keyword) {
         videos = await Video.find({
             title: {
-                $regex: new RegExp(keyword, "i"),
+                $regex: new RegExp(`${keyword}$`, "i"),
             },
-        });
+        }).populate("owner");
     };
     return res.render("search", { pageTitle:"Search", videos });
 };
